lambda_function/requirements.txt
pandas==2.1.3
boto3==1.33.8
botocore==1.33.8
awslambdaric==2.0.8
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0